Fixed issue with Area2Ds and GUI elements not registering the mouse correctly
It turns out this is an issue caused by using integer scaling, not
fractional scaling. This has now been changed (as well as just resizing
the window as a whole so it fits in the editor better). However this
causes fonts to appear gross, so we'll have to figure out a better way
to do this...