Syntax highlighting should work with comments now
The solution was to include comments as tokens, so
CommentTokenizerModule will now actually *return* the comment as a
token. When/if we actually want to *parse* GDScript, we'll want to
filter out those comment tokens, which shouldn't affect the token
stream's syntactic meaning at all. But yeah, we just handle comments
just like any other token now, no weirdness necessary.